Make ktx_off_t type signed as it is supposed to be.

+++ lib/stream.h
@@ -52,7 +52,7 @@
 #if defined(_MSC_VER) && defined(_WIN64)
   typedef unsigned __int64 ktx_off_t;
 #else
-  typedef   size_t ktx_off_t;
+  typedef   off_t ktx_off_t;
 #endif
 typedef struct ktxMem ktxMem;
 typedef struct ktxStream ktxStream;
